refactor: separate handlers into modular components (common, text, audio)
Architectural Improvement:

Separation of Concerns:
- _common.py: Shared utilities and logic (232 lines)
  * Background task error handling
  * Error message transmission
  * Turn limit checking
  * Feedback evaluation (extracted from handlers)
  * AI response generation (extracted from handlers)
  * Feedback message transmission (extracted from handlers)
  * Spring 2 utterance save with conditional feedback

- _text_handler.py: USER_TEXT message handler (161 lines)
  * STT-free text input processing
  * Reuses _common functions for feedback/AI

- _audio_handler.py: UTTERANCE_END message handler (192 lines)
  * Audio-based utterance processing with STT
  * Reuses _common functions for feedback/AI

- message_handlers.py: INIT and END_SESSION handlers (47 lines)
  * Session initialization with STT streaming setup
  * Session cleanup and Spring 2 notification

Benefits:
✅ Reduced code duplication (Option 1 feedback logic extracted)
✅ Improved maintainability (changes to feedback logic in one place)
✅ Better testability (each function independently testable)
✅ Clear separation of concerns (common vs. mode-specific)
✅ Lines per file reduced (796 → 47 + 192 + 161 + 391)

Tests: 114+ tests passing (message router, session handlers, messages, etc.)
